move out class definition to simplify entry point
